MLR-229: Updated existing tests.
