+++ src/test/java/gov/usgs/wma/mlrgateway/BaseSpringTest.java
@@ -7,7 +7,7 @@ import java.util.Map;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.util.FileCopyUtils;
 
-import gov.usgs.wma.mlrgateway.service.LegacyWorkflowService;
+import gov.usgs.wma.mlrgateway.service.workflow.LegacyWorkflowService;
 
 public class BaseSpringTest {
 	public static String legacyValidation = "{\"validation_passed_message\": \"Validations Passed\"}";
